news_recommender/requirements.txt
jsonrpclib-pelix
tap_news_utils
